Made changes to MapFragment to improve geojson rendering. Will now only show additions to osm map when zoomed into a certain level. Also remove markers as they looked messy at any zoom level and were unnecessary now that room labelling was visible.
Modified getFillColour to ensure polygons are transparent/semitransparent to improve visibility. This was due to an issue with a large overlay polygon obscuring the map.

Changed campus_floor_0.geojson to a different example file to better represent changes made.

Added MapListener to monitor zoom levels and trigger displaying/hiding additions to map.

Removed local.properties as this should be unique to all our systems, prevented me from loading the sdk on my machine, you guys might have to change this manually on your end but shouldn't have to do it again now it's removed from git. If you're on widnows you might not have to change it at all.

Added a .gitignore file to clean up the repo and avoid any unnecessary uploads.