Added always visible block labels to the map and rendering system so building blocks are now easily identifiable. e.g. A block, B, C etc
Finally managed to get rid of the annoying speech bubbles which kept popping up whenever I tapped an area of the map overlayed by polygons/polylines. All tap events on these will now be consumed.

Trimmed the geojson file down however we will probably need to trim it right down to the basics that we need as I'm already experiencing some performance issues when tested on my android. Only seems to occur when pinching to zoom, not while using built in zoom controls.